refactor(stellar): move token price fetching out of hooks into StellarPriceService
